# internal/httpapi/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProbeHive API",
    "version": "v1",
    "description": "Contract-compatible API for ProbeHive organization bootstrap, local authentication, Monitors, immutable revisions, Runs, Observations, evaluated health, and Incidents."
  },
...
  "paths": {
...
    "/openapi/v1.json": {
      "get": {
        "tags": ["Health"],
        "operationId": "getOpenAPIV1",
        "summary": "Return this OpenAPI document in Development",
        "responses": {
          "200": {
            "description": "The OpenAPI 3.1 document.",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      },
      "head": {
        "tags": ["Health"],
        "operationId": "headOpenAPIV1",
        "summary": "Check OpenAPI availability in Development",
        "responses": {
          "200": { "description": "The OpenAPI document is available." },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/setup/status": {
      "get": {
        "tags": ["Setup"],
        "operationId": "getSetupStatus",
        "summary": "Report whether the first administrator exists",
        "responses": {
          "200": { "$ref": "#/components/responses/SetupStatus" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/setup/admin": {
      "post": {
        "tags": ["Setup"],
        "operationId": "createFirstAdministrator",
        "summary": "Create the first administrator and provision the installation Organization",
        "description": "Creates the first administrator, provisions the installation Organization with slug 'default' and its default Project through the idempotent provisioning use case, then issues a session cookie. A new installation is usable without a separate Organization step.",
        "security": [
          { "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateFirstAdministratorRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "The first administrator and the installation Organization were created and a session cookie was issued.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SetupResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/auth/antiforgery": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "issueAntiforgeryToken",
        "summary": "Issue or rotate a synchronizer antiforgery token",
        "responses": {
          "200": {
            "description": "The request token and required header name. A host-only antiforgery cookie is also set.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AntiforgeryTokenResponse" } } }
          },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "login",
        "summary": "Authenticate a local user and issue a session",
        "security": [
          { "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "The authenticated user. A fresh session cookie is also set.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "logout",
        "summary": "Invalidate the current session",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "responses": {
          "204": { "description": "The session was invalidated and its cookie expired." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "getSession",
        "summary": "Return the current authenticated session",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "The current user session.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations": {
      "get": {
        "tags": ["Organizations"],
        "operationId": "listOrganizations",
        "summary": "List Organizations with their default Projects",
        "description": "Requires any authenticated session. Returns the caller's Organization memberships in creation order with the Organization id as a tie-breaker, and a bare empty array when none exist.",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "The Organizations visible to the caller.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/OrganizationResponse" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      },
      "post": {
        "tags": ["Organizations"],
        "operationId": "provisionOrganization",
        "summary": "Idempotently provision an Organization and default Project",
        "description": "Requires the instance Administrator role. The creator becomes the new Organization's Administrator member in the same transaction. A matching slug and display name replays with 200 and writes no membership; a new Organization returns 201.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrganizationRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Organization" },
          "201": {
            "description": "The Organization and default Project were created.",
            "headers": { "Location": { "$ref": "#/components/headers/Location" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/name": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "put": {
        "tags": ["Organizations"],
        "operationId": "renameOrganization",
        "summary": "Change an Organization's display name",
        "description": "Requires the organization.write permission. The slug is immutable because it is the idempotency key for provisioning; after a rename, provisioning the same slug with the pre-rename display name conflicts and the current name replays.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RenameOrganizationRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Organization" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "tags": ["Organizations"],
        "operationId": "getOrganization",
        "summary": "Return an Organization and its default Project",
        "description": "Requires an Administrator session.",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Organization" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/webhook-integrations": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "tags": ["Integrations"],
        "operationId": "listWebhookIntegrations",
        "summary": "List signed Webhook Integrations",
        "description": "Requires integration.manage. Returns Administrator-only destination configuration and never returns signing secrets or encrypted-secret metadata.",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "The Organization's Webhook Integrations in creation order.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookIntegrationResponse" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      },
      "post": {
        "tags": ["Integrations"],
        "operationId": "createWebhookIntegration",
        "summary": "Create a disabled signed Webhook Integration",
        "description": "Requires integration.manage and antiforgery. Generates an active signing secret, returns it exactly once, and stores only AES-256-GCM ciphertext. The Integration remains disabled because delivery is not implemented by this slice.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateWebhookIntegrationRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "The disabled Integration and its one-time signing secret.",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the one-time signing secret response.",
                "schema": { "type": "string", "const": "no-store" }
              }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWebhookIntegrationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/webhook-integrations/{integrationId}/state": {
      "parameters": [
        { "$ref": "#/components/parameters/OrganizationID" },
        { "$ref": "#/components/parameters/WebhookIntegrationID" }
      ],
      "put": {
        "tags": ["Integrations"],
        "operationId": "changeWebhookIntegrationState",
        "summary": "Enable or disable a signed Webhook Integration",
        "description": "Requires integration.manage and antiforgery. Uses the current Integration version for optimistic concurrency. At most five Integrations may be enabled in one Organization. Newly projected Alerts receive a stable route for every Integration enabled at projection time; no network call or Delivery Attempt is created.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/WebhookIntegrationStateRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "The Integration in the requested state. Repeating its current state with the current version does not advance the version.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookIntegrationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/webhook-integrations/{integrationId}/signing-secrets/prepare": {
      "parameters": [
        { "$ref": "#/components/parameters/OrganizationID" },
        { "$ref": "#/components/parameters/WebhookIntegrationID" }
      ],
      "post": {
        "tags": ["Integrations"],
        "operationId": "prepareWebhookSigningSecret",
        "summary": "Prepare a new Webhook signing secret",
        "description": "Requires integration.manage and antiforgery. Starts a two-phase rotation, keeps the current secret active, and returns the pending secret exactly once. No second rotation may begin until the pending secret is activated and the former secret is retired.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/WebhookIntegrationVersionRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "The advanced Integration and its one-time pending signing secret.",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the one-time signing secret response.",
                "schema": { "type": "string", "const": "no-store" }
              }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PrepareWebhookSigningSecretResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/webhook-integrations/{integrationId}/signing-secrets/activate": {
      "parameters": [
        { "$ref": "#/components/parameters/OrganizationID" },
        { "$ref": "#/components/parameters/WebhookIntegrationID" }
      ],
      "post": {
        "tags": ["Integrations"],
        "operationId": "activateWebhookSigningSecret",
        "summary": "Activate the prepared Webhook signing secret",
        "description": "Requires integration.manage and antiforgery. Makes the single pending secret active for future calls and leaves the former active secret retiring.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/WebhookIntegrationVersionRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "The Integration with its new active secret version.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookIntegrationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/webhook-integrations/{integrationId}/signing-secrets/retire": {
      "parameters": [
        { "$ref": "#/components/parameters/OrganizationID" },
        { "$ref": "#/components/parameters/WebhookIntegrationID" }
      ],
      "post": {
        "tags": ["Integrations"],
        "operationId": "retireWebhookSigningSecret",
        "summary": "Retire the former Webhook signing secret",
        "description": "Requires integration.manage and antiforgery. Removes ciphertext for the single retiring secret while retaining bounded version and timestamp audit metadata.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/WebhookIntegrationVersionRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "The Integration after the former secret is retired.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookIntegrationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/projects/{projectId}/monitors": {
      "parameters": [
        { "$ref": "#/components/parameters/OrganizationID" },
        { "$ref": "#/components/parameters/ProjectID" }
      ],
      "get": {
        "tags": ["Monitors"],
        "operationId": "listMonitors",
        "summary": "List Monitors in creation order",
        "description": "Requires an Administrator session.",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "The scoped Monitors.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/MonitorResponse" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      },
      "post": {
        "tags": ["Monitors"],
        "operationId": "createMonitor",
        "summary": "Create a draft Monitor",
        "description": "Requires an Administrator session.",
        "security": [
          { "sessionCookie": [], "antiforgeryCookie": [], "antiforgeryHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateMonitorRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "The Monitor was created.",
            "headers": { "Location": { "$ref": "#/components/headers/Location" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MonitorResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/projects/{projectId}/monitors/{monitorId}": {
      "parameters": [
        { "$ref": "#/components/parameters/OrganizationID" },
        { "$ref": "#/components/parameters/ProjectID" },
        { "$ref": "#/components/parameters/MonitorID" }
      ],
      "get": {
        "tags": ["Monitors"],
        "operationId": "getMonitor",
        "summary": "Return a fully scoped Monitor",
        "description": "Requires an Administrator session.",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Monitor" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "probehive.session",
        "description": "Opaque server-side session token. Only its SHA-256 hash is persisted."
      },
      "antiforgeryCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "probehive.antiforgery",
        "description": "Host-only selector cookie issued by the antiforgery endpoint."
      },
      "antiforgeryHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ProbeHive-Antiforgery",
        "description": "Synchronizer request token returned by the antiforgery endpoint."
      }
    },
...
    "responses": {
...
      "SetupStatus": {
        "description": "Current setup state.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SetupStatusResponse" } } }
      },
...
      "NotFound": {
        "description": "The route or fully scoped resource was not found.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      },
...
      "MethodNotAllowed": {
        "description": "The route does not allow this method.",
        "headers": {
          "Allow": {
            "description": "Allowed methods.",
            "schema": { "type": "string" }
          }
        },
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      }
    },
    "schemas": {
...
      "CreateFirstAdministratorRequest": {
        "type": "object",
        "description": "Request property names are matched ASCII-case-insensitively; canonical spellings are shown.",
        "x-probehive-property-name-comparison": "ascii-case-insensitive",
        "required": ["email", "displayName", "password"],
        "properties": {
          "email": { "type": ["string", "null"] },
          "displayName": { "type": ["string", "null"] },
          "password": { "type": ["string", "null"] }
        }
      },
...
    }
  }
}